Added new LightBoard class
LightBoard is a central context to hold all the objects/resources, so
the handler classes just need to be passed in a *LightBoard and have
access to all channels,etc.

NOTE: I'm currently passing digitalChannels in twice because
analogChannels doesn't exist yet.  When it does I will replace the
second one.