Unified digital and master channels to one array, fixed master handling
- There's one array of channels now, and all channel initialization,
handling, keyboard processing, etc is now agnostic of digital or analog
- Implemented AnalogOutPin.  Hardware does not support it yet so not
verified to that level
- Channel needs to have the master channel passed in to properly
calculate the effective level.  No real way around that
- Cleaned up debugging
- Created LightBoard class as a master context to hold all channels.
This can be passed into things like KeyboardHandler and SerialHandler so
they have access to the channels.  I expect the CommandProcessor will
also be added there so handlers can pass commands to it.

Current issues:
- Need to debounce momentary somehow.